refactor(money-formatters): remove directory
