+++ src/shared/ui/utils/money-formatters.test.ts
